requirements.txt
Pillow>=10.0.0
pycryptodome>=3.15.0